+++ packages/embedder/lint-staged.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-  '*.{ts,tsx,graphql}': ['biome check --write '],
+  '*.{ts,tsx,graphql}': () => 'biome check --write --staged',
   '**/*.{ts,tsx}': () => 'tsc --noEmit -p tsconfig.json'
 }