+++ packages/server/email/makeSummaryEmailV2.tsx
@@ -11,17 +11,18 @@ import {Text} from '@react-email/text'
 import dayjs from 'dayjs'
 import type {GraphQLResolveInfo} from 'graphql'
 import {PALETTE} from '../../client/styles/paletteV3'
-import logoSVG from '../../client/styles/theme/images/brand/lockup_color_mark_dark_type.svg'
+import logoImg from '../../client/styles/theme/images/brand/parabol_logo_transparent@1X.png'
 import makeAppURL from '../../client/utils/makeAppURL'
 import plural from '../../client/utils/plural'
 import appOrigin from '../appOrigin'
 import type {DataLoaderInstance} from '../dataloader/RootDataLoader'
-import type {GQLContext} from '../graphql/graphql'
+import type {InternalContext} from '../graphql/graphql'
 import {
   getDimensionNames,
   getPokerRowData
 } from '../graphql/mutations/helpers/summaryPage/getPokerTable'
 import {CipherId} from '../utils/CipherId'
+import {convertTipTapToMarkdown} from '../utils/convertTipTapToMarkdown'
 
 const insightBox = {
   marginBottom: '20px'
@@ -69,10 +70,71 @@ const makeFallbackInsights = async (meetingId: string, dataLoader: DataLoaderIns
   }
   return null
 }
+
+const makeTeamPromptFallbackInsights = async (
+  meetingId: string,
+  dataLoader: DataLoaderInstance
+) => {
+  const responses = await dataLoader.get('teamPromptResponsesByMeetingId').load(meetingId)
+  const responseBlock = await Promise.all(
+    responses.map(async (response) => {
+      const {userId, content} = response
+      if (!userId) return null
+      const user = await dataLoader.get('users').loadNonNull(userId)
+      const {preferredName} = user
+      const markdown = convertTipTapToMarkdown(content)
+
+      return (
+        <>
+          <Text style={{...paragraph, lineHeight: '14px'}} key={userId}>
+            {preferredName}
+          </Text>
+          <Markdown children={markdown} />
+        </>
+      )
+    })
+  )
+  return <Section style={insightBox}>{responseBlock}</Section>
+}
+
+const EmailFooter = (props: {meetingName: string}) => {
+  const {meetingName} = props
+
+  const brandSubtitle = {
+    fontSize: '14px',
+    color: PALETTE.SLATE_700,
+    margin: 0
+  }
+
+  const unsubscribeLink = {
+    color: PALETTE.SKY_500,
+    textDecoration: 'underline'
+  }
+  const unsubscribeURL = makeAppURL(appOrigin, '/me/profile')
+  return (
+    <>
+      <Section style={{marginTop: '32px'}}>
+        <Img src={logoImg} alt='Parabol' style={{marginBottom: '8px'}} />
+        <Text style={brandSubtitle}>Collaborative Workflows & Insights</Text>
+      </Section>
+
+      <Section>
+        <Text
+          style={{fontSize: '12px', marginBottom: '0px'}}
+        >{`${meetingName} Workflow Summary`}</Text>
+        <Text style={{fontSize: '12px', marginTop: '0px', lineHeight: '12px'}}>
+          <a href={unsubscribeURL} style={unsubscribeLink}>
+            Unsubscribe from workflow summaries
+          </a>
+        </Text>
+      </Section>
+    </>
+  )
+}
 export const makeSummaryEmailV2 = async (
   meetingId: string,
   pageId: number,
-  context: GQLContext,
+  context: InternalContext,
   info: GraphQLResolveInfo
 ) => {
   const {dataLoader} = context
@@ -85,6 +147,7 @@ export const makeSummaryEmailV2 = async (
     name: meetingName,
     endedAt,
     teamId,
+    agendaItemCount,
     topicCount,
     taskCount,
     reflectionCount,
@@ -132,20 +195,8 @@ export const makeSummaryEmailV2 = async (
     paddingLeft: '12px',
     paddingRight: '12px'
   }
-
-  const brandSubtitle = {
-    fontSize: '14px',
-    color: PALETTE.SLATE_700,
-    margin: 0
-  }
-
-  const unsubscribeLink = {
-    color: PALETTE.SKY_500,
-    textDecoration: 'underline'
-  }
   const pageCode = CipherId.encrypt(pageId)
   const CTAURL = makeAppURL(appOrigin, `/pages/${pageCode}`)
-  const unsubscribeURL = makeAppURL(appOrigin, '/me/profile')
   const endLabel = endTime.format('MMM D, YYYY')
   const title = `${meetingName} Summary - ${endLabel}`
   const participantLabel = `${meetingMembers.length} ${plural(meetingMembers.length, 'Participant')}`
@@ -192,20 +243,7 @@ export const makeSummaryEmailV2 = async (
                 See Retro Insights & Tasks
               </Button>
             </Section>
-
-            <Section style={{marginTop: '32px'}}>
-              <Img src={logoSVG} alt='Parabol' style={{marginBottom: '8px'}} />
-              <Text style={brandSubtitle}>Collaborative Workflows & Insights</Text>
-            </Section>
-
-            <Section>
-              <Text style={{fontSize: '12px', marginBottom: '0px'}}>Retro Workflow Summary</Text>
-              <Text style={{fontSize: '12px', marginTop: '0px', lineHeight: '12px'}}>
-                <a href={unsubscribeURL} style={unsubscribeLink}>
-                  Unsubscribe from workflow summaries
-                </a>
-              </Text>
-            </Section>
+            <EmailFooter meetingName='Retro' />
           </Container>
         </Body>
       </Html>
@@ -281,19 +319,80 @@ export const makeSummaryEmailV2 = async (
                 </tbody>
               </table>
             </Section>
-            <Section style={{marginTop: '32px'}}>
-              <Img src={logoSVG} alt='Parabol' style={{marginBottom: '8px'}} />
-              <Text style={brandSubtitle}>Collaborative Workflows & Insights</Text>
-            </Section>
+            <EmailFooter meetingName='Sprint Poker' />
+          </Container>
+        </Body>
+      </Html>
+    )
+  } else if (meetingType === 'teamPrompt') {
+    const responses = await dataLoader.get('teamPromptResponsesByMeetingId').load(meetingId)
+    const responseCount = responses.length
+    const responseLabel = `${responseCount} ${plural(responseCount || 0, 'response')}`
+    const subHeadingMeta = `${responseLabel}`
+    const {content: insightsMarkdown} = await dataLoader
+      .get('meetingInsightsContent')
+      .load(meetingId)
 
-            <Section>
-              <Text style={{fontSize: '12px', marginBottom: '0px'}}>Retro Workflow Summary</Text>
-              <Text style={{fontSize: '12px', marginTop: '0px', lineHeight: '12px'}}>
-                <a href={unsubscribeURL} style={unsubscribeLink}>
-                  Unsubscribe from workflow summaries
-                </a>
+    const markdownWithoutBullets = insightsMarkdown
+      ? // replace bullets with a newline
+        insightsMarkdown
+          .replace(/^\s*([-*]|\d+\.)\s+/gm, '\n')
+          .replace(/\n{2,}/g, '\n\n')
+          .trim()
+      : null
+    const fallbackInsights = await makeTeamPromptFallbackInsights(meetingId, dataLoader)
+    return () => (
+      <Html>
+        <Head />
+        <Preview>{title}</Preview>
+        <Body style={main}>
+          <Container style={container}>
+            <Section style={{marginBottom: '20px'}}>
+              <Text style={heading}>{title}</Text>
+              <Text style={subheading}>
+                <strong>{subHeading}</strong>
+                <br />
+                {subHeadingMeta}
               </Text>
             </Section>
+            {markdownWithoutBullets && (
+              <Markdown markdownContainerStyles={insightBox} children={markdownWithoutBullets} />
+            )}
+            {fallbackInsights}
+            <Section style={{marginBottom: '32px'}}>
+              <Button style={ctaButton} href={CTAURL}>
+                See Responses in Parabol
+              </Button>
+            </Section>
+            <EmailFooter meetingName='Standup' />
+          </Container>
+        </Body>
+      </Html>
+    )
+  } else if (meetingType === 'action') {
+    const topicLabel = `${agendaItemCount} ${plural(agendaItemCount || 0, 'Agenda Item')}`
+    const taskLabel = `${taskCount} ${plural(taskCount || 0, 'New Task')}`
+    const subHeadingMeta = `${taskLabel} • ${topicLabel}`
+    return () => (
+      <Html>
+        <Head />
+        <Preview>{title}</Preview>
+        <Body style={main}>
+          <Container style={container}>
+            <Section style={{marginBottom: '20px'}}>
+              <Text style={heading}>{title}</Text>
+              <Text style={subheading}>
+                <strong>{subHeading}</strong>
+                <br />
+                {subHeadingMeta}
+              </Text>
+            </Section>
+            <Section style={{marginBottom: '32px'}}>
+              <Button style={ctaButton} href={CTAURL}>
+                See New Tasks
+              </Button>
+            </Section>
+            <EmailFooter meetingName={'Check-in'} />
           </Container>
         </Body>
       </Html>